requirements.txt
beautifulsoup4
discord.py==1.0.1
requests
urllib3
